chore: Use Java assertions
This is mostly a moot commit used to fill submission requirements. There
is no real need to have assertions as this point, because assertions
should be avoided anyway (and instead handled with exceptions).

The code at this point contains no complex logic, and even if it did, it
is unlikely that assertions would be needed. The assertions written here
check for conditions that would have resulted in an RT error anyway.

The only added value is having more descriptive error messages.